requirements.txt
# YouTube Automation Factory — core deps
supabase>=2.0.0
python-dotenv>=1.0.0
